backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pandas==2.2.3
yfinance==0.2.51
apscheduler==3.11.0
pydantic==2.10.4
python-dotenv==1.0.1
